Extract watchdog functionality into separate module
