+++ chalice/cli/reloader.py
@@ -24,20 +24,17 @@ worker process).
 
 """
 import subprocess
-import threading
 import logging
 import copy
 import sys
 
-import watchdog.observers
-from watchdog.events import FileSystemEventHandler
-from watchdog.events import FileSystemEvent  # noqa
-from typing import MutableMapping, Type, Callable, Optional  # noqa
+from typing import MutableMapping, Type, Callable  # noqa
 
-from chalice.local import LocalDevServer  # noqa
+from chalice.cli.filewatch import RESTART_REQUEST_RC
+from chalice.cli.filewatch.eventbased import WatchdogWorkerProcess
+from chalice.local import LocalDevServer, HTTPServerThread  # noqa
 
 
-RESTART_REQUEST_RC = 3
 LOGGER = logging.getLogger(__name__)
 
 
@@ -47,61 +44,14 @@ def start_parent_process(env):
     process.main()
 
 
-class Restarter(FileSystemEventHandler):
-
-    def __init__(self, restart_event):
-        # type: (threading.Event) -> None
-        # The reason we're using threading
-        self.restart_event = restart_event
-
-    def on_any_event(self, event):
-        # type: (FileSystemEvent) -> None
-        # If we modify a file we'll get a FileModifiedEvent
-        # as well as a DirectoryModifiedEvent.
-        # We only care about reloading is a file is modified.
-        if event.is_directory:
-            return
-        self()
-
-    def __call__(self):
-        # type: () -> None
-        self.restart_event.set()
-
-
 def start_worker_process(server_factory, root_dir):
     # type: (Callable[[], LocalDevServer], str) -> int
     t = HTTPServerThread(server_factory)
-    worker = WorkerProcess(t)
+    worker = WatchdogWorkerProcess(t)
     LOGGER.debug("Starting worker...")
     rc = worker.main(root_dir)
     LOGGER.info("Restarting local dev server.")
     return rc
-
-
-class HTTPServerThread(threading.Thread):
-    """Thread that manages starting/stopping local HTTP server.
-
-    This is a small wrapper around a normal threading.Thread except
-    that it adds shutdown capability of the HTTP server, which is
-    not part of the normal threading.Thread interface.
-
-    """
-    def __init__(self, server_factory):
-        # type: (Callable[[], LocalDevServer]) -> None
-        threading.Thread.__init__(self)
-        self._server_factory = server_factory
-        self._server = None  # type: Optional[LocalDevServer]
-        self.daemon = True
-
-    def run(self):
-        # type: () -> None
-        self._server = self._server_factory()
-        self._server.serve_forever()
-
-    def shutdown(self):
-        # type: () -> None
-        if self._server is not None:
-            self._server.shutdown()
 
 
 class ParentProcess(object):
@@ -127,52 +77,6 @@ class ParentProcess(object):
             except KeyboardInterrupt:
                 process.terminate()
                 raise
-
-
-class WorkerProcess(object):
-    """Worker that runs the chalice dev server."""
-    def __init__(self, http_thread):
-        # type: (HTTPServerThread) -> None
-        self._http_thread = http_thread
-        self._restart_event = threading.Event()
-        self._restart_callback = Restarter(self._restart_event)
-
-    def main(self, project_dir, timeout=None):
-        # type: (str, Optional[int]) -> int
-        self._http_thread.start()
-        self._start_file_watcher(project_dir)
-        if self._restart_event.wait(timeout):
-            self._http_thread.shutdown()
-            return RESTART_REQUEST_RC
-        return 0
-
-    def _start_file_watcher(self, project_dir):
-        # type: (str) -> None
-        watcher = WatchdogFileWatcher()
-        watcher.watch_for_file_changes(
-            project_dir, self._restart_callback)
-
-
-class FileWatcher(object):
-    """Base class for watching files for changes."""
-
-    def watch_for_file_changes(self, root_dir, callback):
-        # type: (str, Callable[[], None]) -> None
-        """Recursively watch directory for changes.
-
-        When a changed file is detected, the provided callback
-        is immediately invoked and the current scan stops.
-
-        """
-        raise NotImplementedError("watch_for_file_changes")
-
-
-class WatchdogFileWatcher(FileWatcher):
-    def watch_for_file_changes(self, root_dir, callback):
-        # type: (str, Callable[[], None]) -> None
-        observer = watchdog.observers.Observer()
-        observer.schedule(callback, root_dir, recursive=True)
-        observer.start()
 
 
 def run_with_reloader(server_factory, env, root_dir):
